Get Started page -matching design, globalized colors, top-right auth alerts and rename to login
